add formatter module
