+++ docs/modules/helpers.ts.md
@@ -1,6 +1,6 @@
 ---
 title: helpers.ts
-nav_order: 1
+nav_order: 2
 parent: Modules
 ---
