+++ docs/modules/parser.ts.md
@@ -1,6 +1,6 @@
 ---
 title: parser.ts
-nav_order: 3
+nav_order: 4
 parent: Modules
 ---
