+++ docs/modules/route.ts.md
@@ -1,6 +1,6 @@
 ---
 title: route.ts
-nav_order: 4
+nav_order: 5
 parent: Modules
 ---
