+++ src/index.ts
@@ -1,7 +1,6 @@
 /**
  * @since 0.4.0
  */
-import { Contravariant1 } from 'fp-ts/lib/Contravariant'
 import * as E from 'fp-ts/lib/Either'
 import { identity } from 'fp-ts/lib/function'
 import * as O from 'fp-ts/lib/Option'
@@ -9,6 +8,7 @@ import * as O from 'fp-ts/lib/Option'
 import { pipe } from 'fp-ts/lib/pipeable'
 import { failure, Int, string, success, Type } from 'io-ts'
 
+import { contramap, format, Formatter, formatter } from './formatter'
 import { RowLacks } from './helpers'
 import {
   alt,
@@ -102,67 +102,30 @@ export {
    */
   map
 }
+
+export {
+  /**
+   * @category formatters
+   * @since 0.4.0
+   */
+  Formatter,
+  /**
+   * @category formatters
+   * @since 0.4.0
+   */
+  format,
+  /**
+   * @category formatters
+   * @since 0.5.1
+   */
+  formatter,
+  /**
+   * @category formatters
+   * @since 0.5.1
+   */
+  contramap
+}
 // ---
-
-/**
- * @category formatters
- * @since 0.4.0
- */
-export class Formatter<A> {
-  /**
-   * @since 0.4.0
-   */
-  readonly _A!: A
-  constructor(readonly run: (r: Route, a: A) => Route) {}
-  /**
-   * @since 0.4.0
-   */
-  contramap<B>(f: (b: B) => A): Formatter<B> {
-    return new Formatter((r, b) => this.run(r, f(b)))
-  }
-  /**
-   * @since 0.4.0
-   */
-  then<B>(that: Formatter<B> & Formatter<RowLacks<B, keyof A>>): Formatter<A & B> {
-    return new Formatter((r, ab) => that.run(this.run(r, ab), ab))
-  }
-}
-
-/**
- * @category formatters
- * @since 0.4.0
- */
-export function format<A>(formatter: Formatter<A>, a: A, encode: boolean = true): string {
-  return formatter.run(Route.empty, a).toString(encode)
-}
-
-declare module 'fp-ts/lib/HKT' {
-  interface URItoKind<A> {
-    'fp-ts-routing/Formatter': Formatter<A>
-  }
-}
-
-const FORMATTER_URI = 'fp-ts-routing/Formatter'
-
-type FORMATTER_URI = typeof FORMATTER_URI
-
-/**
- * @category formatters
- * @since 0.5.1
- */
-export const formatter: Contravariant1<FORMATTER_URI> = {
-  URI: FORMATTER_URI,
-  contramap: (fa, f) => fa.contramap(f)
-}
-
-/**
- * @category formatters
- * @since 0.5.1
- */
-export const contramap =
-  <A, B>(f: (b: B) => A) =>
-  (fa: Formatter<A>): Formatter<B> =>
-    formatter.contramap(fa, f)
 
 /**
  * @category matchers
